feat: implement add command and refactor folders
