+++ src/operations/index.js
@@ -1,0 +1,4 @@
+export {addNewFile} from './addNewFile.js';
+export {readDir} from './readDir.js';
+export {upOperation} from './up.js';
+export {cdOperation} from './cd.js'
